requirements.txt
click==8.1.3
feedgen==0.9.0
Flask==2.2.3
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
Werkzeug==2.2.3
zipp==3.15.0
